+++ gamemode/huds/hud_lambda_player.lua
@@ -3,42 +3,185 @@ local border_w = 5
 local matHover = Material("gui/ps_hover.png", "nocull")
 local boxHover = GWEN.CreateTextureBorder(border, border, 64 - border * 2, 64 - border * 2, border_w, border_w, border_w, border_w, matHover)
 
+local PANEL = {}
+
 cvars.AddChangeCallback("lambda_playermdl", function()
     net.Start("LambdaPlayerSettingsChanged")
     net.SendToServer()
 end, "LambdaPlayerModelChanged")
 
-local PANEL = {}
+cvars.AddChangeCallback("lambda_playermdl_skin", function()
+    net.Start("LambdaPlayerSettingsChanged")
+    net.SendToServer()
+end, "LambdaPlayerModelSkinChanged")
+
+cvars.AddChangeCallback("lambda_playermdl_bodygroup", function()
+    net.Start("LambdaPlayerSettingsChanged")
+    net.SendToServer()
+end, "LambdaPlayerModelBGChanged")
 
 function PANEL:Init()
-    local mdls = GAMEMODE:GetAvailablePlayerModels()
+    local sheetPanel = self:Add("DPropertySheet")
+    sheetPanel:Dock(FILL)
 
-    for name, v in pairs(mdls) do
-        local icon = vgui.Create("SpawnIcon")
+    local mdlPanel = sheetPanel:Add("DPanel")
+    mdlPanel:Dock(FILL)
 
-        if istable(v) and #v > 1 then
-            icon:SetModel(v[2])
-        elseif istable(v) then
-            icon:SetModel(v[1])
-        else
-            icon:SetModel(v)
-        end
+    local searchBar = mdlPanel:Add("DTextEntry")
+    searchBar:Dock(TOP)
+    searchBar:DockMargin(0, 0, 0, 8)
+    searchBar:SetUpdateOnType(true)
+    searchBar:SetPlaceholderText("Search for model")
+    searchBar:SetPlaceholderColor(Color(140, 140, 140, 255))
 
-        icon:SetSize(64, 64)
-        icon:SetTooltip(name)
+    local mdlListPanel = mdlPanel:Add("DPanelSelect")
+    mdlListPanel:Dock(FILL)
 
-        icon.PaintOver = function(this, w, h)
+    local modelTbl = GAMEMODE:GetAvailablePlayerModels()
+    for name, v in pairs(modelTbl) do
+        local item = mdlListPanel:Add("SpawnIcon")
+        item:SetModel(v)
+        item:SetSize(64, 64)
+        item:SetTooltip(name)
+        item.plymdl = name
+        item.mdlPath = player_manager.TranslatePlayerModel(name)
+        item.PaintOver = function(this, w, h)
             if this.OverlayFade > 0 then
                 boxHover(0, 0, w, h, Color(255, 255, 255, this.OverlayFade))
             end
-
             this:DrawSelections()
         end
+        mdlListPanel:AddPanel(item, {lambda_playermdl = name})
+    end
 
-        self:AddPanel(icon, {
-            lambda_playermdl = name
-        })
+    searchBar.OnValueChange = function(s, str)
+        for i, pnl in pairs(mdlListPanel:GetItems()) do
+            if not pnl.plymdl:find(str, 1, true) and not pnl.mdlPath:find(str, 1, true) then
+                pnl:SetVisible(false)
+            else
+                pnl:SetVisible(true)
+            end
+        end
+        mdlListPanel:InvalidateLayout()
+    end
+
+    sheetPanel:AddSheet("MODEL", mdlPanel)
+
+    local bgPanel = sheetPanel:Add("DPanel")
+
+    local bgList = bgPanel:Add("DPanelList")
+    bgList:Dock(FILL)
+    bgList:EnableVerticalScrollbar(true)
+
+    local nobgLabel = bgPanel:Add("DLabel")
+    nobgLabel:Dock(TOP)
+    nobgLabel:SetText("NO OPTIONS FOR THIS MODEL")
+    nobgLabel:SetTextColor(Color(255, 255, 255, 255))
+    nobgLabel:SetFont("TargetIDSmall")
+    nobgLabel:SizeToContents()
+    nobgLabel:DockMargin(0, 5, 0, 0)
+    nobgLabel:SetContentAlignment(5)
+    nobgLabel:SetVisible(false)
+
+    local bgTab = sheetPanel:AddSheet("BODYGROUPS", bgPanel)
+
+    local function HighlightTab(state)
+        if state then
+            bgTab.Tab:SetTextColor(Color(248, 128, 0, 255))
+            nobgLabel:SetVisible(false)
+        else
+            bgTab.Tab:SetTextColor(Color(140, 140, 140, 255))
+            nobgLabel:SetVisible(true)
+        end
+    end
+
+    local function SetMdlChanges(pnl, val)
+        if pnl.type == "skin" then
+            lambda_playermdl_skin:SetString(math.Round(val))
+        end
+
+        if pnl.type == "bg" then
+            local str = string.Explode(" ", lambda_playermdl_bodygroup:GetString())
+            if #str < pnl.n + 1 then
+                for i = 1, pnl.n + 1 do
+                    str[i] = str[i] or 0 end
+            end
+            str[pnl.n + 1] = math.Round(val)
+            lambda_playermdl_bodygroup:SetString(table.concat(str, " "))
+        end
+    end
+
+    local function RebuildBgPnl()
+        bgList:Clear()
+        HighlightTab(false)
+        -- Slight delay to make sure model is set on entity
+        timer.Simple(0.1, function()
+            local ply = LocalPlayer()
+            local mdlStr = player_manager.TranslatePlayerModel(lambda_playermdl:GetString())
+            local numSkins = NumModelSkins(mdlStr) - 1
+            if numSkins > 0 then
+                local slider = vgui.Create("DNumSlider")
+                slider:Dock(TOP)
+                slider:DockMargin(20, 0, 0, 0)
+                slider:SetText("SKIN")
+                slider:SetSkin("Lambda")
+                slider:SetTall(30)
+                slider:SetDecimals(0)
+                slider:SetMax(numSkins)
+                slider:SetValue(lambda_playermdl_skin:GetString())
+                slider:GetTextArea():SetFont("TargetIDSmall")
+                slider:GetTextArea():SetTextColor(Color(255, 255, 255, 255))
+                slider.Label:SetFont("TargetIDSmall")
+                slider.Label:SetTextColor(Color(255, 255, 255, 255))
+                slider.type = "skin"
+                slider.OnValueChanged = SetMdlChanges
+
+                bgList:AddItem(slider)
+                HighlightTab(true)
+            end
+
+            local bgroups = string.Explode(" ", lambda_playermdl_bodygroup:GetString())
+            for k = 0, ply:GetNumBodyGroups() - 1 do
+                if ply:GetBodygroupCount(k) <= 1 then continue end
+
+                local bgsldr = vgui.Create("DNumSlider")
+                bgsldr:Dock(TOP)
+                bgsldr:DockMargin(20, 0, 0, 0)
+                bgsldr:SetText(string.upper(ply:GetBodygroupName(k)))
+                bgsldr:SetSkin("Lambda")
+                bgsldr:SetTall(30)
+                bgsldr:SetDecimals(0)
+                bgsldr:SetMax(ply:GetBodygroupCount(k) - 1)
+                bgsldr:SetValue(bgroups[k + 1] or 0)
+                bgsldr:GetTextArea():SetFont("TargetIDSmall")
+                bgsldr:GetTextArea():SetTextColor(Color(255, 255, 255, 255))
+                bgsldr.Label:SetFont("TargetIDSmall")
+                bgsldr.Label:SetTextColor(Color(255, 255, 255, 255))
+                bgsldr.type = "bg"
+                bgsldr.n = k
+                bgsldr.OnValueChanged = SetMdlChanges
+
+                bgList:AddItem(bgsldr)
+                HighlightTab(true)
+            end
+        end)
+    end
+
+    function RebuildPanel()
+        RebuildBgPnl()
+    end
+
+    RebuildPanel()
+
+    function mdlListPanel:OnActivePanelChanged(old, new)
+        if old != new then
+            lambda_playermdl_skin:SetString("0")
+            lambda_playermdl_bodygroup:SetString("0")
+        end
+        timer.Simple(0.1, function()
+            RebuildPanel()
+        end)
     end
 end
 
-vgui.Register("LambdaPlayerPanel", PANEL, "DPanelSelect")
+vgui.Register("LambdaPlayerPanel", PANEL, "DPanel")